backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.5
python-dotenv==1.0.1
openai==1.54.0
pypdf==4.3.1
sentence-transformers==3.0.1
chromadb==0.5.5